requirements.txt
Django==1.5.1
django-tastypie==0.9.15
djorm-ext-core==0.4.1
djorm-ext-expressions==0.4.3
djorm-ext-pgarray==0.4.2
mimeparse==0.1.3
psycopg2==2.5
python-dateutil==2.1
six==1.3.0
wsgiref==0.1.2
